Somewhat working Targed Crawl
At this point, we can repeatedly query a single node however there are
a couple of bugs...

We currently do not search the entire space because we error out due to
build ups of callbacks to a single node. This may mean we can only be
waiting on a single callback per node at a time.... need to investigate
further.

Support has been added for logging both to a file and to the console so
we can trace back what occured in previous runs, all logs are currently
going to the same file, howevever updating what file is written to is as
simple as updating the myLogger file.